AuthorityServiceImple 추가

+++ src/main/java/com/nhnacademy/taskApi/repository/ProjectStatusRepository.java
@@ -3,5 +3,5 @@ package com.nhnacademy.taskApi.repository;
 import com.nhnacademy.taskApi.domain.ProjectStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectStatusRepository extends JpaRepository<ProjectStatus, Integer> {
+public interface ProjectStatusRepository extends JpaRepository<ProjectStatus, Long> {
 }
